refactor(querykeys): shift to ts and place it in constants
