+++ src/constants/queryKeys.ts
