+++ src/hooks/directoryHooks/useCreateDirectory.ts
@@ -1,7 +1,7 @@
 import { AxiosError } from "axios"
 import { useMutation, UseMutationResult, useQueryClient } from "react-query"
 
-import { DIR_CONTENT_KEY, RESOURCE_ROOM_NAME_KEY } from "hooks/queryKeys"
+import { DIR_CONTENT_KEY, RESOURCE_ROOM_NAME_KEY } from "constants/queryKeys"
 
 import * as DirectoryService from "services/DirectoryService/index"
