+++ src/hooks/directoryHooks/useCreateDirectoryHook.jsx
@@ -1,9 +1,10 @@
 import { useContext } from "react"
 import { useMutation, useQueryClient } from "react-query"
 
+import { DIR_CONTENT_KEY } from "constants/queryKeys"
+
 import { ServicesContext } from "contexts/ServicesContext"
 
-import { DIR_CONTENT_KEY } from "hooks/queryKeys"
 import useRedirectHook from "hooks/useRedirectHook"
 
 import { useErrorToast } from "utils/toasts"
