+++ src/hooks/moveHooks/useMoveHook.jsx
@@ -1,9 +1,9 @@
 import { useContext } from "react"
 import { useMutation, useQueryClient } from "react-query"
 
-import { ServicesContext } from "contexts/ServicesContext"
+import { DIR_CONTENT_KEY } from "constants/queryKeys"
 
-import { DIR_CONTENT_KEY } from "hooks/queryKeys"
+import { ServicesContext } from "contexts/ServicesContext"
 
 import { useSuccessToast, useErrorToast } from "utils/toasts"
