+++ src/hooks/settingsHooks/useGetSettings.ts
@@ -2,6 +2,8 @@ import _, { zip } from "lodash"
 import { useQuery, UseQueryResult } from "react-query"
 import type { StringKeyOf } from "type-fest"
 
+import { SETTINGS_CONTENT_KEY } from "constants/queryKeys"
+
 import * as SettingsService from "services/SettingsService"
 
 import {
@@ -10,8 +12,6 @@ import {
   SiteSettings,
   SiteSocialMediaSettings,
 } from "types/settings"
-
-import { SETTINGS_CONTENT_KEY } from "../queryKeys"
 
 import { BE_TO_FE } from "./constants"
